fix: move bookmarks to a Store
