+++ apps/ui/src/stores/spaces.ts
@@ -5,8 +5,17 @@ import { NetworkID } from '@/types';
 export const useSpacesStore = defineStore('spaces', () => {
   const metaStore = useMetaStore();
 
-  const { loading, loaded, networksMap, spaces, spacesMap, hasMoreSpaces, fetch, fetchMore } =
-    useSpaces();
+  const {
+    loading,
+    loaded,
+    networksMap,
+    spaces,
+    spacesMap,
+    hasMoreSpaces,
+    fetch,
+    fetchMore,
+    getSpaces
+  } = useSpaces();
 
   async function fetchSpace(spaceId: string, networkId: NetworkID) {
     await metaStore.fetchBlock(networkId);
@@ -31,6 +40,7 @@ export const useSpacesStore = defineStore('spaces', () => {
     hasMoreSpaces,
     fetch,
     fetchMore,
-    fetchSpace
+    fetchSpace,
+    getSpaces
   };
 });
